requirements.txt
fastapi
uvicorn[standard]
python-dotenv

groq

google-api-python-client
google-auth-httplib2
google-auth-oauthlib

gspread
google-auth
